mempool: fix stack handler dequeue
[ upstream commit e09ff22d538e13ead91a5caa2d136e8e065e5ed1 ]

The return value of the stack handler is wrong: it should be 0 on
success, not the number of objects dequeued.

This could lead to memory leaks depending on how the caller checks the
return value (ret < 0 or ret != 0). This was also breaking autotests
with debug enabled, because the debug cookies are only updated when the
function returns 0, so the cookies were not updated, leading to
an abort().

Fixes: 295a530b0844 ("mempool: add stack mempool handler")

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>