net/ixgbe: fix received packets number for ARM NEON
[ upstream commit 989a84050542ba1eca247ec4bf4b98a1173f6aef ]

Vector PMD will check 4 descs in one time, but the statuses are not
consistent because the memory allocated for RX descriptors is cacheable
huagepage.
This patch is to calculate the number of received packets by scann DD bit
sequentially, and stops when meeting the first packet with DD bit unset.

Fixes: b20971b6cca0 ("net/ixgbe: implement vector driver for ARM")

Signed-off-by: Jianbo Liu <jianbo.liu@linaro.org>