net/ixgbe: fix default MAC setting
[ upstream commit c53e2c6f9253365f544df87014989e267274155c ]

Pool 0 is not PF, it's VF 0. So the MAC is set for VF 0
but not PF.
The code introduced a weird issue. In the scenario PF + VF,
when only starting PF, the default PF MAC address is working.
But after starting a VF, the default PF MAC address becomes
the VF's address.

Use the pool which is not occupied by VFs for PF to fix it.

Fixes: 8164fe82846b ("ixgbe: add default mac address modifier")

Signed-off-by: Wenzhuo Lu <wenzhuo.lu@intel.com>