net/bonding: fix number of bonding Tx/Rx queues
[ upstream commit acfb51e2fe96c27a042248ee32930cbd78f36cd7 ]

This patch fixes the maximum number of Tx an Rx queues supported by a
bonding device return by the rte_eth_dev_info_get function.

The bonding device now calculates the maximum number of supported Tx
and Rx queues based on the slaves bound to the bonded device, with the
minimum values of Tx and Rx queues from the device slaves being the
bonded devices maximum, as each slave must be able to support the same
number of Tx and Rx queues.

Fixes: 2efb58cbab6e ("bond: new link bonding library")

Signed-off-by: Declan Doherty <declan.doherty@intel.com>