test/memzone: fix memory leak
[ upstream commit 71330483a19303e5001ea70af80672cbe6cd881b ]

This patch fixes the memory leaks in memzone_autotest. Those memory leaks
lead to failures in tests from the same testing group due to out of memory
problems.  With introduction of rte_memzone_free() it is now possible to
free the memzone.  Fix uses this API call to make a clean after each test
case.

Fixes: ff909fe21f0a ("mem: introduce memzone freeing")

Signed-off-by: Radoslaw Biernacki <radoslaw.biernacki@linaro.org>