Added support for RTE_FLOW_ACTION_TYPE_DROP. A drop filter is created dropping all packets matching the filter. The packet drop is made by the SmartNIC.
Added support for RTE_FLOW_ACTION_TYPE_PORT_ID.
A forward filter is created forwarding all packets matching the filter to the DPDK port `id`. Forwarded packets are not sent to the APP.
The packet forward is made by the SmartNIC. The DPDK  port `id` must be on the same SmartNIC as the RX port.

Added support for RTE_FLOW_ACTION_TYPE_PHY_PORT.
Works in the same ways as RTE_FLOW_ACTION_TYPE_PORT_ID, but the port `index` must be the physical port number of the SmartNIC.
For a 4 ported SmartNIC the port `index` must be 0 to 3.

tx_offload_capa is set to DEV_TX_OFFLOAD_MULTI_SEGS