eal: check string parameter lengths
[ upstream commit 2383d8e909275b072145c7addf826e9c8edb6232 ]

When specifying parameters such as hugefile prefix from the
command-line, it is possibly to supply an empty string. This may
lead to various problems: for example, if hugefile prefix is
empty, the runtime config path construction may end up
looking like "/var/run/dpdk//_config", which will technically
work, but is wrong and places files in the wrong place.

To fix it, check lengths of such user-specified parameters for
hugefile prefix, as well as hugepage dir and user-specified
mbuf pool ops string.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>