net/bonding: fix link status
[ upstream commit d7bce0058a241a4d9b03e295bd74e6957b040d1e ]

Copying the link properties of the first slave added may copy an
invalid link status. The speed and duplex of the slave may not
be known at this time. Delay setting the properties until the
first slave reports as link up. Note that we are still ignoring
an error from link_properties_valid. For some bonding modes,
802.3ad, we should not activate the slave if it does not have
matching link properties.

Fixes: a45b288ef21a ("bond: support link status polling")

Signed-off-by: Chas Williams <chas3@att.com>