net/nfp: fix misuse of strlcpy
Current strlcpy function is doing the wrong thing and as a consequence
the firmware does not find the symbol requested precluding the right
NFP initialization.

Using strncpy is safe here since the symbol length can never be longer
than the buffer size where the firmware will get the symbol to work
with. However, newer compilers do not allow to have the strncpy using
the source length as the third parameter, so this patch uses instead a
memcpy call with a previous memset for cleaning up the buffer to be
used by the firmware.

Fixes: a5d659c2d03f ("net/nfp: replace strncpy by strlcpy")

Signed-off-by: Alejandro Lucero <alejandro.lucero@netronome.com>