mem: warn user when running without NUMA support
[ upstream commit 23d5455517568966ca8cac216941128ed3d3a455 ]

Running in non-legacy mode on a NUMA-enabled system without libnuma
is unsupported, so explicitly print out a warning when trying to
do so.

Running in legacy mode without libnuma is still supported whether or
not we are running with libnuma support enabled, so also fix init to
allow that scenario.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>