Revert "net/mlx5: fix instruction hotspot on replenishing Rx buffer"
This reverts commit 40454190463981cc1c382a6470cd64970a4025c3.

Reverting this commit as it is being reverted and replaced with
a different fix in master dpdk due to a slight performance regression
in some circumstances. At a later date when the new fix is applied in
master dpdk and validated, it may be backported to the stable branch.

Discussed in:
https://mails.dpdk.org/archives/stable/2019-March/013567.html

Cc: Yongseok Koh <yskoh@mellanox.com>
Signed-off-by: Kevin Traynor <ktraynor@redhat.com>