net/bnx2x: fix race for periodic flags
[ upstream commit c19eafccb22e1b1ca52439a936908cfd6fd12ff1 ]

The periodic callout function and the interrupt handler both modify the
periodic flags. There is a possible race condition when an application
is going through dev_stop()/dev_start() and an interrupt handler is
invoked. We also need to ensure that periodic polling is not invoked in
interrupt context.

This patch handles such case by using separate variable to check for
interrupt context. Also, atomically load and store the periodic flag
value.

Fixes: 0f6ebeee2402 ("net/bnx2x: fix call to link handling periodic function")

Signed-off-by: Rasesh Mody <rmody@marvell.com>