kni: fix kernel 5.4 build - num_online_cpus
num_online_cpus was redefined [1] to now be a static inline function.
Due to that the check we had to redefine num_online_cpus as needed
triggers as false-positive and breaks like:

/usr/src/dpdk-rte-kni-18.11.2/ethtool/ixgbe/kcompat.h:239:27:
error: ‘smp_num_cpus’ undeclared (first use in this function)
  239 | #define num_online_cpus() smp_num_cpus

That redefine isn't needed for ages, and with >5.4 we can just rely on
it. From the kernel it already is either a working function or defined
as 1UL in case of non SMP.

Not applying to master branch due to kni ethtool being removed in [2]

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=0c09ab96
[2]: https://git.dpdk.org/dpdk/commit/?id=ea6b39b

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Acked-by: Luca Boccassi <bluca@debian.org>